[TypeDeclaration] Split ParamTypeByMethodCallTypeRector into Object, Scalar and Array rules
Extract shared engine into AbstractParamTypeByMethodCallTypeRector; each concrete rule
owns a disjoint param type group via isMatchingParamType():

- ObjectParamTypeByMethodCallTypeRector - object types (incl. nullable)
- ScalarParamTypeByMethodCallTypeRector - scalar types (string/int/float/bool)
- ArrayParamTypeByMethodCallTypeRector  - array types (incl. nullable)
- ParamTypeByMethodCallTypeRector       - remaining compound (cross-group unions, iterable, callable)

Types are null-stripped first, so nullable variants route to their group.
