+++ composer.json
@@ -100,7 +100,7 @@
         "files": [
             "tests/debug_functions.php",
             "rules-tests/Transform/Rector/FuncCall/FuncCallToMethodCallRector/Source/some_view_function.php",
-            "rules-tests/TypeDeclaration/Rector/ClassMethod/ParamTypeByMethodCallTypeRector/Source/FunctionTyped.php",
+            "rules-tests/TypeDeclaration/Rector/ClassMethod/ScalarParamTypeByMethodCallTypeRector/Source/FunctionTyped.php",
             "rules-tests/TypeDeclaration/Rector/StmtsAwareInterface/SafeDeclareStrictTypesRector/Source/functions.php",
             "rules-tests/Php70/Rector/ClassMethod/Php4ConstructorRector/Source/ParentClass.php"
         ]
