+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ArrayParamTypeByMethodCallTypeRector/Fixture/skip_scalar_type.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ArrayParamTypeByMethodCallTypeRector\Fixture;
+
+final class SkipScalarType
+{
+    public function go($value)
+    {
+        $this->use($value);
+    }
+
+    private function use(string $value): void
+    {
+    }
+}
