+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ObjectParamTypeByMethodCallTypeRector/Fixture/nullable_object_param.php.inc
@@ -1,0 +1,39 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ObjectParamTypeByMethodCallTypeRector\Fixture;
+
+use stdClass;
+
+final class NullableObjectParam
+{
+    public function go($value)
+    {
+        $this->use($value);
+    }
+
+    private function use(?stdClass $value): void
+    {
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ObjectParamTypeByMethodCallTypeRector\Fixture;
+
+use stdClass;
+
+final class NullableObjectParam
+{
+    public function go(?\stdClass $value)
+    {
+        $this->use($value);
+    }
+
+    private function use(?stdClass $value): void
+    {
+    }
+}
+
+?>
