+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ObjectParamTypeByMethodCallTypeRector/Fixture/skip_array_type.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ObjectParamTypeByMethodCallTypeRector\Fixture;
+
+final class SkipArrayType
+{
+    public function go($value)
+    {
+        $this->use($value);
+    }
+
+    private function use(array $value): void
+    {
+    }
+}
