+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ObjectParamTypeByMethodCallTypeRector/Fixture/skip_scalar_type.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ObjectParamTypeByMethodCallTypeRector\Fixture;
+
+final class SkipScalarType
+{
+    public function go($value)
+    {
+        $this->use($value);
+    }
+
+    private function use(string $value): void
+    {
+    }
+}
