--- rules-tests/TypeDeclaration/Rector/ClassMethod/ParamTypeByMethodCallTypeRector/Fixture/avoid_nested_call.php.inc
@@ -1,45 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Fixture;
-
-use Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Source\SomeTypedService;
-
-final class AvoidNestedCall
-{
-    public function __construct(
-        private SomeTypedService $someTypedService
-    ) {
-    }
-
-    public function go($value)
-    {
-        $result = function ($value) {
-            $this->someTypedService->run($value);
-        };
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Fixture;
-
-use Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Source\SomeTypedService;
-
-final class AvoidNestedCall
-{
-    public function __construct(
-        private SomeTypedService $someTypedService
-    ) {
-    }
-
-    public function go($value)
-    {
-        $result = function (string $value) {
-            $this->someTypedService->run($value);
-        };
-    }
-}
-
-?>
