--- rules-tests/TypeDeclaration/Rector/ClassMethod/ParamTypeByMethodCallTypeRector/Fixture/default_null.php.inc
@@ -1,41 +1,0 @@
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Fixture;
-
-use Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Source\SomeTypedService;
-
-final class DefaultNull
-{
-    public function __construct(
-        private SomeTypedService $someTypedService
-    ) {
-    }
-
-    public function go($value)
-    {
-        $this->someTypedService->withDefaultNull($value);
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Fixture;
-
-use Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Source\SomeTypedService;
-
-final class DefaultNull
-{
-    public function __construct(
-        private SomeTypedService $someTypedService
-    ) {
-    }
-
-    public function go(?string $value)
-    {
-        $this->someTypedService->withDefaultNull($value);
-    }
-}
-
-?>
