+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ParamTypeByMethodCallTypeRector/Fixture/skip_pure_array.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector\Fixture;
+
+final class SkipPureArray
+{
+    public function go($value)
+    {
+        $this->use($value);
+    }
+
+    private function use(array $value): void
+    {
+    }
+}
