+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ScalarParamTypeByMethodCallTypeRector/Fixture/func_call_elsewhere.php.inc
@@ -1,0 +1,31 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ScalarParamTypeByMethodCallTypeRector\Fixture;
+
+use function Rector\Tests\TypeDeclaration\Rector\ClassMethod\ScalarParamTypeByMethodCallTypeRector\Source\function_typed;
+
+final class FuncCallElsewhere
+{
+    public function go($anotherValue, $someValue)
+    {
+        function_typed($someValue, $anotherValue);
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ScalarParamTypeByMethodCallTypeRector\Fixture;
+
+use function Rector\Tests\TypeDeclaration\Rector\ClassMethod\ScalarParamTypeByMethodCallTypeRector\Source\function_typed;
+
+final class FuncCallElsewhere
+{
+    public function go(int|bool $anotherValue, $someValue)
+    {
+        function_typed($someValue, $anotherValue);
+    }
+}
+
+?>
