+++ rules-tests/TypeDeclaration/Rector/ClassMethod/ScalarParamTypeByMethodCallTypeRector/Fixture/skip_object_type.php.inc
@@ -1,0 +1,17 @@
+<?php
+
+namespace Rector\Tests\TypeDeclaration\Rector\ClassMethod\ScalarParamTypeByMethodCallTypeRector\Fixture;
+
+use stdClass;
+
+final class SkipObjectType
+{
+    public function go($value)
+    {
+        $this->use($value);
+    }
+
+    private function use(stdClass $value): void
+    {
+    }
+}
