+++ rules/TypeDeclaration/Rector/ClassMethod/ScalarParamTypeByMethodCallTypeRector.php
@@ -1,0 +1,75 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\TypeDeclaration\Rector\ClassMethod;
+
+use PHPStan\Type\Type;
+use PHPStan\Type\TypeCombinator;
+use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
+use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
+
+/**
+ * Handles the scalar (string/int/float/bool) param type group.
+ *
+ * @see \Rector\Tests\TypeDeclaration\Rector\ClassMethod\ScalarParamTypeByMethodCallTypeRector\ScalarParamTypeByMethodCallTypeRectorTest
+ */
+final class ScalarParamTypeByMethodCallTypeRector extends AbstractParamTypeByMethodCallTypeRector
+{
+    public function getRuleDefinition(): RuleDefinition
+    {
+        return new RuleDefinition('Change scalar param type based on passed method call type', [
+            new CodeSample(
+                <<<'CODE_SAMPLE'
+class SomeTypedService
+{
+    public function run(string $name)
+    {
+    }
+}
+
+final class UseDependency
+{
+    public function __construct(
+        private SomeTypedService $someTypedService
+    ) {
+    }
+
+    public function go($value)
+    {
+        $this->someTypedService->run($value);
+    }
+}
+CODE_SAMPLE
+                ,
+                <<<'CODE_SAMPLE'
+class SomeTypedService
+{
+    public function run(string $name)
+    {
+    }
+}
+
+final class UseDependency
+{
+    public function __construct(
+        private SomeTypedService $someTypedService
+    ) {
+    }
+
+    public function go(string $value)
+    {
+        $this->someTypedService->run($value);
+    }
+}
+CODE_SAMPLE
+            ),
+        ]);
+    }
+
+    protected function isMatchingParamType(Type $type): bool
+    {
+        return TypeCombinator::removeNull($type)->isScalar()
+            ->yes();
+    }
+}
